feat: implement per-app language settings
