+++ app/src/main/res/resources.properties
@@ -1,0 +1,1 @@
+unqualifiedResLocale=en
